Successfully made a single image!

+++ analysis.py
@@ -122,7 +122,7 @@ class Slice(DedalusData):
             
     def read_data(self):
         self.times = np.array([])
-        self.writes = np.array([])
+        self.writes = np.array([], dtype=np.int)
         
         N = 1
         for filename in self.files:
@@ -141,7 +141,7 @@ class Slice(DedalusData):
             self.z = f['scales']['z']['1.0'][:]
 
             self.times = np.append(self.times, f['scales']['sim_time'][:])
-            self.writes = np.append(self.writes, f['scales']['write_number'][:])
+            self.writes = np.append(self.writes, int(f['scales']['write_number'][:]))
             f.close()
             
         for key in self.keys:
